requirements.txt
bentoml==1.4.19
category_encoders==2.8.1
fastapi==0.116.1
FLAML==2.3.5
matplotlib==3.10.5
mlflow==2.17.0
mlflow_skinny==2.17.0
pandas==2.3.1
prometheus_client==0.21.1
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.11.7
python-dotenv==1.1.1
scikit_learn==1.7.1
starlette==0.47.2
xgboost==3.0.4
